CATCH THE DIAMOND/requirements.txt
pygame
PyOpenGL 
PyOpenGL_accelerate